+++ src/main/java/net/quantrax/citybuild/module/support/SupportManager.java
@@ -1,6 +1,5 @@
-package net.quantrax.citybuild.support;
+package net.quantrax.citybuild.module.support;
 
-import de.derioo.manager.CommandFramework;
 import io.papermc.paper.event.player.AsyncChatEvent;
 import lombok.Getter;
 import net.dv8tion.jda.api.entities.channel.ChannelType;
@@ -10,11 +9,10 @@ import net.kyori.adventure.audience.Audience;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.MiniMessage;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.quantrax.citybuild.support.chat.Chat;
-import net.quantrax.citybuild.support.commands.SupportCommand;
-import net.quantrax.citybuild.support.discordbot.DiscordBot;
-import net.quantrax.citybuild.support.player.QueueMember;
-import net.quantrax.citybuild.support.player.Supporter;
+import net.quantrax.citybuild.module.support.chat.Chat;
+import net.quantrax.citybuild.module.support.discordbot.DiscordBot;
+import net.quantrax.citybuild.module.support.player.QueueMember;
+import net.quantrax.citybuild.module.support.player.Supporter;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
