+++ src/main/java/net/quantrax/citybuild/module/support/chat/Chat.java
@@ -1,21 +1,15 @@
-package net.quantrax.citybuild.support.chat;
+package net.quantrax.citybuild.module.support.chat;
 
 import com.google.common.base.Preconditions;
-import io.papermc.paper.chat.ChatRenderer;
-import io.papermc.paper.event.player.AsyncChatEvent;
 import lombok.Getter;
 import lombok.Setter;
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import net.kyori.adventure.audience.Audience;
-import net.kyori.adventure.audience.ForwardingAudience;
-import net.kyori.adventure.text.minimessage.MiniMessage;
-import net.quantrax.citybuild.support.SupportManager;
-import net.quantrax.citybuild.support.player.QueueMember;
-import net.quantrax.citybuild.support.player.Supporter;
+import net.quantrax.citybuild.module.support.SupportManager;
+import net.quantrax.citybuild.module.support.player.QueueMember;
+import net.quantrax.citybuild.module.support.player.Supporter;
 import org.bukkit.Bukkit;
-import org.bukkit.Server;
 import org.bukkit.entity.Player;
-import org.w3c.dom.stylesheets.LinkStyle;
 
 import java.util.ArrayList;
 import java.util.List;
