+++ src/main/java/net/quantrax/citybuild/module/support/discordbot/DiscordBot.java
@@ -1,11 +1,11 @@
-package net.quantrax.citybuild.support.discordbot;
+package net.quantrax.citybuild.module.support.discordbot;
 
 import lombok.Getter;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.requests.GatewayIntent;
-import net.quantrax.citybuild.support.SupportManager;
+import net.quantrax.citybuild.module.support.SupportManager;
 
 @Getter
 public class DiscordBot {
