+++ src/main/java/net/quantrax/citybuild/module/support/player/Supporter.java
@@ -1,7 +1,7 @@
-package net.quantrax.citybuild.support.player;
+package net.quantrax.citybuild.module.support.player;
 
 import lombok.Getter;
-import net.quantrax.citybuild.support.chat.Chat;
+import net.quantrax.citybuild.module.support.chat.Chat;
 
 import java.util.ArrayList;
 import java.util.List;
